Don't do comm allocations for arrays that will be serially initialized
The main reason to use the comm layer allocation interface is so that we can do
chapel initialization before registering the array, so that we can get good
numa affinity. However there are non-trivial overheads associates with
allocating using the comm layer API, so if we're just going to do serial
initialization, we might as well just allocate through the normal API
allocation API, which will still give us the right affinity for the current
thread.

Note that under ugni, we'll if the extensible heap has to be extended, we'll
still do a comm layer allocation and all that entails, but we'll only do that
once and jemalloc will keep the memory, so the cost should be amortized over
the lifetime of the program.

TODO, we probably don't want to do this if the user has disabled the extensible
heap by setting CHPL_RT_MAX_HEAP_SIZE