Add more heuristics to the array init calculation
If we're running serially, or we're only going to create a single chunk in
range iteration (because the runningTaskCounter doesn't merit it) do serial
initialization. This is motivated by wanting to avoid comm layer allocation for
ISx. We know the array will be initialized serially, so we want to skip the
expensive comm layer allocation. On master we still would have heuristically
chosen parallel initialization and the range iterator would have determined not
to create any tasks. Here we just duplicate some of that logic so that we can
detect serial init at the point where we'll call the allocation interface.