Track which arrays are forced to use dynamic heap extensions
The previous work didn't quite work. When we skipped comm allocation for
"serial" arrays we still mistakenly freed them with the comm_free routine. This
is because under the covers the dynamic heap extension code still uses the comm
alloc so as far as the comm layer knows it did allocate the array.

Here just make DefaultRectangular track whether is forced arrays to be serially
initialized. An alternative (and probably less error-prone) method would be to
keep an array or hash-table to explicit track which arrays allocations were
done through comm-alloc, and which ones were done with chpl_mem_alloc